Fix Windows C4244/C4267/C4334/C4996 warnings
Generative_Quantum_Machine_Learning_Base.cpp:
  - Line 91/838/874: size_t→int via static_cast
  - Lines 239/324/385: 1<<qbit_num → size_t(1)<<qbit_num (C4334)
  - Lines 641/678: memset 0.0 → 0 (C4244 double→int)
  - Line 629: suppress C4996 unsafe fopen with #pragma warning(suppress)

N_Qubit_Decomposition_Tree_Search.cpp:
  - Line 281: static_cast<int8_t>(topology.size()) for std::fill (C4267)
  - Line 785: 1u<<m → size_t(1)<<m for vector size (C4334 x2)

density_matrix.cpp:
  - Lines 557/665: static_cast<int>(*.size()) (C4267)

apply_large_kernel_to_input_AVX.cpp (float32 kernel only):
  - Lines 4333-4334: std::fill(..., 0.0) → 0.0f (C4244)
  - Lines 4357-4358: static_cast<float> for double→float assignment
  - Line 7467: double results[8] → float results[8] (C4244)

density_matrix/bindings.cpp:
  - Lines 36/124/164: static_cast<int>(buf.shape[0]) (C4244 ssize_t→int)

Also fix macOS calc_one_qubit_u3/parameters_for_calc_one_qubit override
warnings in H.h, S.h, SDG.h, SX.h, SXdg.h, T.h, Tdg.h, X.h, Y.h, Z.h,
RX.h, RZ.h, RZ_P.h, U1.h, U2.h

+++ squander/src-cpp/decomposition/N_Qubit_Decomposition_Tree_Search.cpp
@@ -278,7 +278,7 @@ void generate_insertions_recursive(
 
     if (depth == num_cnot) {
         matrix_base<int8_t> limits = matrix_base<int8_t>(1, curpath.size()+num_cnot);
-        std::fill(limits.data, limits.data + limits.size(), topology.size());
+        std::fill(limits.data, limits.data + limits.size(), static_cast<int8_t>(topology.size()));
         GrayCodeCNOT out(limits);
 
         int j = 0, k = 0;
@@ -782,7 +782,7 @@ bool contains_topological_subsequence(
 
     // reachable[S] = whether subset S of small nodes can be matched
     // after scanning some prefix of big
-    std::vector<char> reachable(1u << m, 0), next_reachable(1u << m, 0);
+    std::vector<char> reachable(size_t(1) << m, 0), next_reachable(size_t(1) << m, 0);
     reachable[0] = 1;
 
     for (int i = 0; i < bigpath.size(); i++) {
